+++ frontend/lost/src/actions/auth.ts
@@ -2,7 +2,7 @@ import axios from 'axios'
 import { jwtDecode } from 'jwt-decode'
 import { isEmpty } from 'lodash'
 import { API_URL } from '../lost_settings'
-import { DecodedToken } from './auth/auth_types'
+import { DecodedToken } from '../types/auth_types'
 
 const refreshToken = async () => {
   const jwt: string | null = localStorage.getItem('refreshToken')
