+++ frontend/lost/src/types/model.ts
